zram: introduce automatic writeback
zram will now writeback idle pages to the backing device automatically.

If a page in zram is not accessed in wb_start_mins, it'll be written to
the backing device when the device screen goes off.

To prevent aggressive writebacks, time accounting is only done when
touch events are present. Take this into consideration as wb_start_mins
doesn't represent system uptime. This prevents aggressively writing back pages
when the device is sleeping and slowdowns from handling page faults when
the device wakes up (e.g., using the phone after waking up in the morning).

Time accounting is now done with sched_clock() instead of ktime as we can
tolerate scheduler's timer drifts.

Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>
Signed-off-by: rk134 <rahul@aospa.co>