docs/vm/damon: remove broken reference
DAMON commits listing script had typos, so it was missing three DAMON
documentation fixes.  For the reason, the three fixes were not
backported on this tree.  This and following two commits backport those.

[1] https://git.kernel.org/sj/damon-hack/c/721ba248f9d5

This patch (of 3):

commit 876d0aac2e3af10fbaf1c7a814840c71e470dc5c upstream.

Building DAMON documents warns for a reference to nonexisting doc, as
below:

    $ time make htmldocs
    [...]
    Documentation/vm/damon/index.rst:24: WARNING: toctree contains reference to nonexisting document 'vm/damon/plans'

This fixes the warning by removing the wrong reference.

Link: https://lkml.kernel.org/r/20210917123958.3819-4-sj@kernel.org
Signed-off-by: SeongJae Park <sjpark@amazon.de>
Cc: Jonathan Corbet <corbet@lwn.net>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: SeongJae Park <sjpark@amazon.com>
Signed-off-by: rk134 <rahul@aospa.co>